Separate translator functionallity out of tree route stack into extending implementation to remove hard dependency
