+++ library/Zend/Mvc/composer.json
@@ -26,6 +26,9 @@
         "zendframework/zend-text": "self.version",
         "zendframework/zend-view": "self.version"
     },
+    "suggest": {
+        "zendframework/zend-i18n": "Zend\\I18n component for translatable segments"
+    },
     "extra": {
         "branch-alias": {
             "dev-master": "2.1-dev",
